% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/wrappers.R
\name{f_agL}
\alias{f_agL}
\title{Optimization function}
\usage{
f_agL(
  y,
  design,
  cov_mat,
  basis_num,
  weights,
  lambda_1,
  lambda_2,
  Ps,
  initN,
  centered = FALSE,
  active_dim = 1e-06,
  max_iter = 1000L,
  standardnorm = FALSE
)

f_agL(
  y,
  design,
  cov_mat,
  basis_num,
  weights,
  lambda_1,
  lambda_2,
  Ps,
  initN,
  centered = FALSE,
  active_dim = 1e-06,
  max_iter = 1000L,
  standardnorm = FALSE
)
}
\arguments{
\item{y}{Response vector.}

\item{design}{Design matrix.}

\item{cov_mat}{List of covariance matrices.}

\item{basis_num}{Number of basis functions per variable.}

\item{weights}{Weights for group lasso.}

\item{lambda_1}{Tuning parameter for group lasso.}

\item{lambda_2}{Vector of tuning parameters for roughness penalty.}

\item{Ps}{Block diagonal penalty matrix.}

\item{initN}{Initial value (optional).}

\item{centered}{Whether to center input.}

\item{active_dim}{Threshold for eigenvalue truncation.}

\item{max_iter}{Maximum number of iterations.}

\item{standardnorm}{Whether to use standard L2 norm.}
}
\value{
Estimated coefficients.
}
\description{
Optimization function

Functional adaptive group lasso estimator
}
