% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/wrappers.R
\name{matpower_1}
\alias{matpower_1}
\title{Power function for matrix}
\usage{
matpower_1(a, alpha, active_dim = 1e-06)

matpower_1(a, alpha, active_dim = 1e-06)
}
\arguments{
\item{a}{A symmetric matrix.}

\item{alpha}{Power exponent.}

\item{active_dim}{Threshold to ignore small eigenvalues.}
}
\value{
a matrix powered by alpha

A matrix raised to the power alpha.
}
\description{
Computes the matrix power \eqn{A^\alpha} for a symmetric matrix.
}
