% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAGLasso_base_functions.R
\name{numeric_integ}
\alias{numeric_integ}
\title{numerical integration}
\usage{
numeric_integ(x, tt)
}
\arguments{
\item{x}{vector of f(t_i), tt: vector of t_i}

\item{tt}{integration of f on the range of tt}
}
\value{
integration of f on the range of tt
}
\description{
numerical integration
}
