% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/wrappers.R
\name{soft_thresh}
\alias{soft_thresh}
\title{Soft thresh function}
\usage{
soft_thresh(x)

soft_thresh(x)
}
\arguments{
\item{x}{A numeric value.}
}
\value{
\code{x * I(x > 0)}.

\code{x} if positive, 0 otherwise.
}
\description{
Soft thresh function

Soft-threshold function
}
