% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{f_agL}
\alias{f_agL}
\title{Optimization function}
\usage{
f_agL(
  y,
  design,
  cov_mat,
  basis_num,
  weights,
  lambda_1,
  lambda_2,
  Ps,
  initN,
  centered = FALSE,
  active_dim = 1e-06,
  max_iter = 1000L,
  standardnorm = FALSE
)
}
\arguments{
\item{y}{a vector of response variable.}

\item{design}{a matrix of coefficients for explanatory functions.}

\item{cov_mat}{a list of coordinate representations for covariance operators.}

\item{basis_num}{a vector of the numbers of basis functions to represent explanatory functions.}

\item{weights}{a vector of weights for penalties. If all elements of it is set to be one, it is the case for the non-adaptive group lasso.}

\item{lambda_1}{a hyperparameter for the group lasso penalty.}

\item{lambda_2}{a vector of hyperparameters for the roughness penalties.}

\item{Ps}{a diagonal block matrix of second level gram matrices.}

\item{initN}{a vector of initial value. It defaults to be NULL.}

\item{centered}{a Boolean value. If true, y and design are centered. Otherwise, both y and design are not centered.}

\item{active_dim}{a threshold for inverse of matrix.}

\item{max_iter}{the maximum number of iterations allowed.}

\item{standardnorm}{If set to 'TRUE', the standard L2 norm is employed instead of our gamma norm. Note that it is just for the simulation study.}
}
\description{
Optimization function
}
