% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fpca_cpp}
\alias{fpca_cpp}
\title{Funtional principal component analysis}
\usage{
fpca_cpp(coefs, GB)
}
\arguments{
\item{coefs}{list object with \code{coefs[[j]] = [x^j_{1:n}]} : coordinates of j-th explanatory functions.
Hence, \code{dim( coefs[[j]] ) = c(m_j, n)}, where m_j is the number of basis and n is the number of observations.}

\item{GB}{list object with \code{GB[[j]]} is the j-th Gram matrix such as \code{fourierpen(basis, 0) or bsplinepen(basis, 0)}.}
}
\value{
A list with the following components:
\itemize{
\item \code{pred}: ist object with \code{pred[[j]] = pca scores of j-th explanatory functions, dim(pred[[j]]) = (n, m_j)}.
\item \code{eval}: list object with\code{eval[[j]]} = eigenvalues of j-th covariance operator.
\item \code{GB}: list object with \code{GB[[j]]} = Gram Matrix for j-th explanatory functions.
\item \code{mat}: list object with \code{mat[[j]] = [I]_B^C, operator changing basis system from the original(bspline or fourier)} to the eigenfunctions.
\item \code{Sigma}:  list object with \code{Sigma[[j]]} is the coordinate representation of j-th covariance operator, whose basis system is eigenfunctions.
\item \code{non_centered}: non_centered design matrix based on eigenfunctions.
}
@export
}
\description{
Funtional principal component analysis
}
