% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matpower_1}
\alias{matpower_1}
\title{Power function for matrix}
\usage{
matpower_1(a, alpha, active_dim = 1e-06)
}
\arguments{
\item{a}{matrix to be powered.}

\item{alpha}{a number by which the given matrix is powered.}

\item{active_dim}{a num}
}
\value{
a matrix powered by alpha
}
\description{
Power function for matrix
}
